Stop namespacing components
