+++ app/components/alert_component.html.erb
