+++ app/components/also_available_component.html.erb
