+++ app/components/code_snippet_modal_component.html.erb
