+++ app/components/document/sidebar_component.html.erb
