+++ app/components/document_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class DocumentComponent < Geoblacklight::DocumentComponent
+end
