--- app/components/earthworks/document/sidebar_component.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Earthworks
-  # Render the sidebar on the show view
-  module Document
-    class SidebarComponent < Geoblacklight::Document::SidebarComponent
-      def components
-        [
-          Geoblacklight::LoginLinkComponent.new(document: document),
-          Geoblacklight::StaticMapComponent.new(document: document),
-          Earthworks::DownloadLinksComponent.new(document: document),
-          Earthworks::AlsoAvailableComponent.new(document: document)
-        ]
-      end
-    end
-  end
-end
