--- app/components/earthworks/document_component.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module Earthworks
-  class DocumentComponent < Geoblacklight::DocumentComponent
-    def classes
-      [
-        @classes,
-        helpers.render_document_class(@document),
-        'document',
-        'col-lg-6',
-        ("document-position-#{@counter}" if @counter)
-      ].compact.flatten
-    end
-  end
-end
