--- app/components/earthworks/server_applied_params_component.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module Earthworks
-  class ServerAppliedParamsComponent < Blacklight::SearchContext::ServerAppliedParamsComponent
-  end
-end
