+++ app/components/header_icons_component.html.erb
