+++ app/components/recently_added_list.html.erb
