+++ app/components/search_result_component.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class SearchResultComponent < Geoblacklight::SearchResultComponent
+  def classes
+    [
+      @classes,
+      helpers.render_document_class(@document),
+      'document',
+      'p-2',
+      'mt-0',
+      ("document-position-#{@counter}" if @counter)
+    ].compact.flatten
+  end
+end
