+++ app/components/server_applied_params_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class ServerAppliedParamsComponent < Blacklight::SearchContext::ServerAppliedParamsComponent
+end
