+++ app/views/catalog/code_snippet.html.erb
@@ -1,1 +1,1 @@
-<%= render Earthworks::CodeSnippetModalComponent.new(@docs.first) %>
+<%= render CodeSnippetModalComponent.new(@docs.first) %>
