+++ app/views/layouts/catalog_result.html.erb
@@ -5,7 +5,7 @@
   <section class="<%= show_sidebar_classes %>">
     <%= content_for(:sidebar) %>
   </section>
-  
+
   <section class="<%= show_content_classes %>">
     <%= yield %>
   </section>
